fix(兼容): kv_score 的 torch.mm out_dtype 改输入上浮 fp32
torch.mm(out_dtype=...) 是 torch 2.7+ 的参数，torch 2.5.1 直接
TypeError。compressor/indexer 两处 kv_score 改为输入 .float() 后
mm，等价于 fp32 累加 + fp32 输出。GateLinear 的 out_dtype 是自有
参数且昆仑走 Tier3 F.linear 路径，不受影响。

+++ vllm_kunlun/models/deepseek_v4/attention.py
@@ -387,10 +387,11 @@ class DeepseekV4Attention(nn.Module, AttentionLayerBase, ABC):
             compressor = self.compressor
 
             def compressor_kv_score() -> torch.Tensor:
+                # torch 2.5.1 has no mm(out_dtype=...); upcast inputs to
+                # fp32 instead (same fp32 accumulate + fp32 output).
                 return torch.mm(
-                    hidden_states,
-                    compressor.fused_wkv_wgate.weight.T,
-                    out_dtype=torch.float32,
+                    hidden_states.float(),
+                    compressor.fused_wkv_wgate.weight.T.float(),
                 )
 
             aux_fns[0] = compressor_kv_score
@@ -404,10 +405,10 @@ class DeepseekV4Attention(nn.Module, AttentionLayerBase, ABC):
                 return weights
 
             def indexer_compressor_kv_score() -> torch.Tensor:
+                # Same torch 2.5.1 workaround as compressor_kv_score above.
                 return torch.mm(
-                    hidden_states,
-                    indexer.compressor.fused_wkv_wgate.weight.T,
-                    out_dtype=torch.float32,
+                    hidden_states.float(),
+                    indexer.compressor.fused_wkv_wgate.weight.T.float(),
                 )
 
             aux_fns[1] = indexer_weights_proj
